Bug with endless note or category saving was solved. New SettingsFragment UI.

+++ app/src/main/java/com/terabyte/realmnotes/ui/fragment/SettingsFragment.kt
@@ -1,5 +1,6 @@
 package com.terabyte.realmnotes.ui.fragment
 
+import android.app.AlertDialog
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -10,6 +11,7 @@ import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
+import com.terabyte.realmnotes.R
 import com.terabyte.realmnotes.application.MyApplication
 import com.terabyte.realmnotes.databinding.FragmentSettingsBinding
 import com.terabyte.realmnotes.ui.viewmodel.MainViewModel
@@ -51,16 +53,47 @@ class SettingsFragment : Fragment() {
         }
 
         binding.buttonDeleteAllNotes.setOnClickListener {
-            mainViewModel.deleteAllNotes()
+            showDeleteAllNotesDialog()
         }
 
         binding.buttonDeleteAllCategories.setOnClickListener {
-            mainViewModel.deleteAllCategories()
+            showDeleteAllCategoriesDialog()
         }
     }
 
-    companion object {
+    private fun showDeleteAllNotesDialog() {
+        val dialog = AlertDialog.Builder(requireContext())
+            .setTitle(getString(R.string.dialog_delete_all_notes_title))
+            .setMessage(getString(R.string.dialog_delete_all_notes_msg))
+            .setCancelable(true)
+            .setPositiveButton(getString(R.string.dialog_button_delete)) { _, _ ->
+                mainViewModel.deleteAllNotes()
+            }
+            .setNegativeButton(getString(R.string.cancel)) { dialog, _ ->
+                dialog.dismiss()
+            }
+            .create()
 
+        dialog.show()
+    }
+
+    private fun showDeleteAllCategoriesDialog() {
+        val dialog = AlertDialog.Builder(requireContext())
+            .setTitle(getString(R.string.dialog_delete_all_categories_title))
+            .setMessage(getString(R.string.dialog_delete_all_categories_msg))
+            .setCancelable(true)
+            .setPositiveButton(getString(R.string.dialog_button_delete)) { _, _ ->
+                mainViewModel.deleteAllCategories()
+            }
+            .setNegativeButton(getString(R.string.cancel)) { dialog, _ ->
+                dialog.dismiss()
+            }
+            .create()
+
+        dialog.show()
+    }
+
+    companion object {
         fun newInstance(): SettingsFragment {
             return SettingsFragment()
         }
